+++ src/test/java/com/example/waiter/OtherTests/OrderSchedulerTests.java
@@ -1,4 +1,4 @@
-package com.example.waiter;
+package com.example.waiter.OtherTests;
 
 import com.example.waiter.Entities.Order;
 import com.example.waiter.Repositories.OrderRepository;
